fix(ref-impls): use uint64 chainid encoding to match precompile
Amp-Thread-ID: https://ampcode.com/threads/T-019c6d63-35ca-736d-90ae-6a4973a51a02

+++ tips/ref-impls/test/ValidatorConfigV2.t.sol
@@ -52,7 +52,7 @@ contract ValidatorConfigV2Test is BaseTest {
     {
         bytes32 message = keccak256(
             abi.encodePacked(
-                block.chainid, address(validatorConfigV2), validatorAddress, ingress, egress
+                uint64(block.chainid), address(validatorConfigV2), validatorAddress, ingress, egress
             )
         );
         return vm.signEd25519(
@@ -72,7 +72,7 @@ contract ValidatorConfigV2Test is BaseTest {
     {
         bytes32 message = keccak256(
             abi.encodePacked(
-                block.chainid, address(validatorConfigV2), validatorAddress, ingress, egress
+                uint64(block.chainid), address(validatorConfigV2), validatorAddress, ingress, egress
             )
         );
         return vm.signEd25519(
